LPS-54663 completedIndexers is no bigger than indexers.size(), it should be int, not long

+++ modules/portal/portal-search/src/main/java/com/liferay/portal/search/internal/background/task/ReindexBackgroundTaskStatusMessageTranslator.java
@@ -82,7 +82,7 @@ public class ReindexBackgroundTaskStatusMessageTranslator
 			String lastIndexer = GetterUtil.getString(
 				backgroundTaskStatus.getAttribute("lastIndexer"));
 
-			long completedIndexers = GetterUtil.getLong(
+			int completedIndexers = GetterUtil.getInteger(
 				backgroundTaskStatus.getAttribute("completedIndexers"));
 
 			if (Validator.isNull(lastIndexer)) {
